backend/requirements.txt
fastapi
uvicorn
python-multipart
pyyaml
pillow
paddleocr
opencv-python
numpy